+++ apps/uniapp/src/main.ts
@@ -1,3 +1,4 @@
+import './polyfill';
 import { createSSRApp } from 'vue';
 import {
   NodeEnv,
@@ -10,12 +11,17 @@ import {
   notify
 } from '@vtj/uni-app';
 import * as VueI18n from 'vue-i18n';
-
+import { createUniAppAxiosAdapter } from '@uni-helper/axios-adapter';
 import App from './App.vue';
 import { name } from '../package.json';
 import { createModules } from './modules';
 
-const adapter = createAdapter({ loading, notify });
+const adapter = createAdapter({
+  loading,
+  notify,
+  axiosAdapter: createUniAppAxiosAdapter()
+});
+
 const service = new LocalService(createServiceRequest(notify));
 const modules = createModules();