+++ create-vtj/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.22](https://gitee.com/newgateway/vtj/compare/create-vtj@0.12.21...create-vtj@0.12.22) (2026-06-11)
+
+
+### Features
+
+* **uniapp:** 集成 axios 适配器并添加 URLSearchParams 补丁 ([59ac3a2](https://gitee.com/newgateway/vtj/commits/59ac3a2b1c55b26538cdfa6605d8927cba544526))
+
+
+
+
+
+## [0.12.21](https://gitee.com/newgateway/vtj/compare/create-vtj@0.12.20...create-vtj@0.12.21) (2026-06-11)
+
+
+### Features
+
+* **uniapp:** add polyfills for FormData and Blob; fix useAssets splice ([14e4ab4](https://gitee.com/newgateway/vtj/commits/14e4ab4671b190426d10e3172cc42675623bf012))
+
+
+
+
+
 ## [0.12.20](https://gitee.com/newgateway/vtj/compare/create-vtj@0.12.19...create-vtj@0.12.20) (2026-06-09)